refrain from including some of the docs
Debian handles copyright information in a standardized way, see the
debian/copyright file, which is installed into the package's doc
directory.  It is considered a bug to supply redundant copyright &
licensing information in the deb.

Also, instructions on how to install and release pycam don't belong in
the package.

This patch was contributed by Sebastian Kuzminsky - thanks!


git-svn-id: https://pycam.svn.sourceforge.net/svnroot/pycam/trunk@265 bbaffbd6-741e-11dd-a85d-61de82d9cad9