reduce unnecessary moves to safety height for zigzag patterns
Previously the "TinySidewaysMoveFilter" was used to detect small gaps within
a toolpath that can be removed safely. But the above filter only worked under
the specific conditions that the z levels of both adjacent points are the same.
The new implementation adds another connection line to the motion grid (before
collision tests). Thus now all kind of connections (independent of their length
and position) are possible.