switch from flake8 to pyflakes
This is not an ideal change, but i don't know of a better way to make
things work in Travis.

The python-flake8 package in Debian Jessie and Stretch provides a __main__
suitable for the "python -m flake8" use that the 'make check-style'
target used.

Ubuntu Trusty (which is what Travis uses) does *not* provide __main__,
instead it provides a /usr/bin/flake8 program that does the same thing.

The /usr/bin/flake8 program is included in Jessie's python-flake8 but
not in Stretch's.

Stretch has a different deb named 'flake8' which provides this binary,
but it only checks Python3.

To summarize:

    "python -m flake8" works on Jessie and Stretch but not Trusty

    "flake8" works on Jessie and Trusty but not Stretch

I found this "pyflakes" program that's available with a consistent
interface on Jessie, Stretch, Precise, and Trusty.  I'm not sure how
the coverage of pyflakes vs flake8 compares.

I'm not making this change because i think pyflakes is better, i'm only
making it because it lets me move forward with Travis.