handle older versions of dpkg-parsechangelog
The Makefile's update-deb-changelog target needs to determine
the most recent released version.

Before this commit, it used to call "dpkg-parsechangelog -S version",
but that fails on the Travis server, because the dpkg-parsechangelog
version is too old.

This commit changes it to run dpkg-parsechangelog without any arguments,
and then uses sed to pull the version number out of the output.
This should work on both old and new dpkg-parsechangelog versions.