setup.py: restore packages list
The packages[] list in setup.py was changed in 995cf490 to just say
"pycam", but that makes it not include all the components of the pycam
library in the setup.py dist output (at least on Debian Stretch and on
the Ubuntu 14.04 "Trusty").

This commit restores the original packages list so that all components
of the pycam library are included again.