restore pycam.Geometry.Line.get_line_length()
This function was removed in commit 52b80196 "removed unused
code (reported by 'vulture')", but it is actually used by
pycam.Geometry.Polygon.is_connectable().  vulture just missed it.

You can see the problem that arises when this function is missing by
following this recipe:

    1. Start pycam.
    2. Delete the default model.
    3. Import a 2d model.
    4. Create a new Tool (all default).
    5. Create a new Process, Engraving, select the 2d model from step 3,
       enable radius compensation.
    6. Create a new Bounds (all default).
    7. Create a new Task (all default)>
    8. Click Generate Toolpath.

Before this commit (but after 52b80196), pycam fails to produce a
sensible toolpath.  After this commit, the correct toolpath is generated.