add example model for polygon offsetting
The sample 'rectangle_chamfered_corners.dxf' contains a rectangle with
slightly chamfered corners (exactly the same as in pycam.Test.test_polygon).

The test_polygon works as expected.
The DXF model (with exactly the same dimensions) works only for a few
tool diameters, e.g. 5.6, 7.0 and 8.0.
It fails for most values in between.

Thus the polygon offsetting algorithm still seems to be unstable (but
probably less than before).  Currently the culprit seems to be
"simplify_polygon_intersections" - it fails to detect an intersection
within a polygon (which is supposed to cause a split of the plugin).