Flow: do not remove item from collection via "__del__"
It could happen, that the item was recently removed from a list and
added again (as a new instance with the same ID) later.
In this case the delayed cleanup via "__del__" would cause the same item
to be removed erroneously again (thus the item would be removed from the
list twice).