Fix, clarify and comment the location code for surfacing
Previously the locations for calculating the toolpath for a surface
operation flattened a surface with a slight slope.  The issues was based
on an overly relaxed check for the alignment of points.
Additionally the interpolation of additional tool locations (in case of
non-linear moves) was unevenly spread along the tool path.  Thus
relevant locations (small model features) could be skipped accidentally.